[#46] feat: 수정하기 로직 및 컴포넌트 추가
